Add jQuery to Sample.Presentation
I notice that _Layout.cshtml references a lot of files in the "lib" directory, but the "lib" directory doesn't exist (rather, it didn't exist until I created it in this commit). What's up with that? If the default templates refer to the "lib" directory, surely the "lib" directory should be created automatically?